backend/auth/requirements.txt
bcrypt==4.0.1
PyJWT==2.8.0
psycopg2-binary==2.9.7